+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Allow splitting before `get` and `set` (#462).
 * Add `--indent` to specify leading indent (#464).
 * Ensure collection elements line split separately (#474).
+* Allow redirecting constructors to wrap (#475).
 * Handle index expressions in the middle of call chains.
 * Optimize splitting lines with many rules.